# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  main.CreateProjectPayload:
    properties:
      description:
        type: string
      emoji:
        type: string
      name:
        type: string
      start_date:
        type: string
      target_date:
        type: string
    type: object
...
  main.UpdateProjectPayload:
    properties:
      description:
        type: string
      emoji:
        type: string
      name:
        type: string
      start_date:
        type: string
      target_date:
        type: string
    type: object
...
  store.Project:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      cycles_enabled:
        description: CyclesEnabled is changed only through SetCyclesEnabled, never
          Update.
        type: boolean
      description:
        type: string
      emoji:
        type: string
      id:
        type: integer
      my_role:
        type: string
      name:
        type: string
      start_date:
        type: string
      target_date:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  store.ProjectAccess:
    properties:
      can_manage:
        type: boolean
      is_superadmin:
        type: boolean
      role:
        type: string
    type: object
...
  store.ProjectDetails:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      cycles_enabled:
        description: CyclesEnabled is changed only through SetCyclesEnabled, never
          Update.
        type: boolean
      description:
        type: string
      emoji:
        type: string
      id:
        type: integer
      members:
        items:
          $ref: '#/definitions/store.ProjectMember'
        type: array
      my_access:
        $ref: '#/definitions/store.ProjectAccess'
      my_role:
        type: string
      name:
        type: string
      start_date:
        type: string
      target_date:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  store.ProjectMember:
    properties:
      email:
        type: string
      is_active:
        type: boolean
      joined_at:
        type: string
      name:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for Trackforge application.
  title: Trackforge API
  version: "1.0"
paths:
...
  /projects:
    get:
      description: The superadmin gets every project; everyone else gets the ones
        they belong to. Each project has my_role, the caller's role in it.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Project'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: List the projects the caller can see
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Admins and the superadmin only. The creator becomes the project's
        first admin.
      parameters:
      - description: Project details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateProjectPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Project'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Project admins and the superadmin only.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - projects
    get:
      description: Includes each member's role, and my_access, which tells the caller
        which actions to offer.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.ProjectDetails'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Get a project with its members
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Project admins and the superadmin only. Partial update; omitted
        fields keep their current value.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to change
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateProjectPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Project'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    description: JWT from POST /auth/login. Paste it with the scheme included, e.g.
      "Bearer eyJhbGciOi..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"